# artifacts/api-server/src/middlewares/authMiddleware.ts
import type { NextFunction, Request, Response } from "express";
import {
  clearSession,
  getSession,
  getSessionId,
  updateSession,
  type SessionData,
  type AuthUser,
} from "../lib/auth";
import { refreshSupabaseSession, toAuthUser } from "../lib/supabase-auth";

declare global {
  namespace Express {
    interface User extends AuthUser {}
    interface Request {
      isAuthenticated(): this is AuthedRequest;
      user?: User | undefined;
    }
    interface AuthedRequest {
      user: User;
    }
  }
}

async function refreshIfExpired(
  sid: string,
  session: SessionData,
): Promise<SessionData | null> {
  const now = Math.floor(Date.now() / 1000);
  if (!session.expires_at || now < session.expires_at - 60) return session;
  if (!session.refresh_token) return null;

  try {
    const tokens = await refreshSupabaseSession(session.refresh_token);
    session.access_token = tokens.access_token;
    session.refresh_token = tokens.refresh_token || session.refresh_token;
    session.expires_at = now + tokens.expires_in;
    session.user = toAuthUser(tokens.user);
    await updateSession(sid, session);
    return session;
  } catch {
    return null;
  }
}

export async function authMiddleware(
  req: Request,
  res: Response,
  next: NextFunction,
): Promise<void> {
  req.isAuthenticated = function (this: Request) {
    return this.user != null;
  } as Request["isAuthenticated"];

  const sid = getSessionId(req);
  if (!sid) {
    next();
    return;
  }

  const session = await getSession(sid);
  if (!session?.user?.id) {
    await clearSession(res, sid);
    next();
    return;
  }

  const refreshed = await refreshIfExpired(sid, session);
  if (!refreshed) {
    await clearSession(res, sid);
    next();
    return;
  }

  req.user = refreshed.user;
  next();
}