# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerGenie API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: dashboard
  - name: applications
  - name: recommendations
  - name: interview
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /auth/signup:
    post:
      tags: [Auth]
      operationId: signUp
      summary: Create a Supabase Auth account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "202":
          description: Email confirmation is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthConfirmationResponse"
  /auth/signin:
    post:
      tags: [Auth]
      operationId: signIn
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Session started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshAuthSession
      summary: Refresh the current server session
      responses:
        "200":
          description: Session refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/signout:
    post:
      tags: [Auth]
      operationId: signOut
      summary: Sign out and revoke the Supabase session
      responses:
        "200":
          description: Session cleared.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: List applications
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{applicationId}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get an application workspace
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{applicationId}/recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get explainable CV recommendations
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /recommendations/{recommendationId}:
    patch:
      operationId: updateRecommendation
      tags: [recommendations]
      summary: Accept or reject a recommendation
      parameters:
        - name: recommendationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationUpdate"
      responses:
        "200":
          description: Updated recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
  /applications/{applicationId}/interview:
    post:
      operationId: startInterview
      tags: [interview]
      summary: Start a role-specific mock interview
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "201":
          description: Interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{interviewId}/answer:
    post:
      operationId: submitInterviewAnswer
      tags: [interview]
      summary: Submit an answer for evaluation
      parameters:
        - name: interviewId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewAnswerInput"
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewEvaluation"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token using the Bearer scheme.
      schema: { type: string }
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema: { type: string }
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email: { type: ["string", "null"], format: email }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        profileImageUrl: { type: ["string", "null"] }
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        firstName: { type: string }
        lastName: { type: string }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
    AuthConfirmationResponse:
      type: object
      required: [requiresEmailConfirmation, message]
      properties:
        requiresEmailConfirmation: { type: boolean, enum: [true] }
        message: { type: string }
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code: { type: string, minLength: 1 }
        code_verifier: { type: string, minLength: 1 }
        redirect_uri: { type: string, minLength: 1, format: uri }
        state: { type: string, minLength: 1 }
        nonce: { type: string, minLength: 1 }
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token: { type: string }
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    DashboardSummary:
      type: object
      required: [firstName, applications, interviews, offers, readiness, recentApplications]
      properties:
        firstName: { type: string }
        applications: { type: integer }
        interviews: { type: integer }
        offers: { type: integer }
        readiness: { type: integer }
        recentApplications:
          type: array
          items: { $ref: "#/components/schemas/Application" }
    ApplicationInput:
      type: object
      required: [jobTitle, companyName]
      properties:
        jobTitle: { type: string, minLength: 1 }
        companyName: { type: string, minLength: 1 }
        location: { type: string }
        sourceUrl: { type: ["string", "null"] }
        description: { type: string }
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [saved, cv_tailored, applied, interview, offer, rejected, withdrawn]
    Application:
      type: object
      required: [id, jobTitle, companyName, location, status, matchScore, cvVersion, updatedAt]
      properties:
        id: { type: string }
        jobTitle: { type: string }
        companyName: { type: string }
        location: { type: string }
        status:
          type: string
          enum: [saved, cv_tailored, applied, interview, offer, rejected, withdrawn]
        matchScore: { type: integer }
        cvVersion: { type: ["string", "null"] }
        appliedDate: { type: ["string", "null"] }
        updatedAt: { type: string }
        initials: { type: string }
        accent: { type: string }
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          required: [summary, strengths, gaps, scoreBreakdown, jobDescription, resumeName, acceptedCount, rejectedCount]
          properties:
            summary: { type: string }
            strengths: { type: array, items: { type: string } }
            gaps: { type: array, items: { type: string } }
            scoreBreakdown:
              type: object
              additionalProperties: { type: integer }
            jobDescription: { type: string }
            resumeName: { type: string }
            acceptedCount: { type: integer }
            rejectedCount: { type: integer }
    Recommendation:
      type: object
      required: [id, section, originalText, suggestedText, reason, evidence, alignment, risk, status]
      properties:
        id: { type: string }
        section: { type: string }
        originalText: { type: string }
        suggestedText: { type: string }
        reason: { type: string }
        evidence: { type: array, items: { type: string } }
        alignment: { type: array, items: { type: string } }
        risk:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, accepted, rejected]
    RecommendationUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, accepted, rejected]
    InterviewSession:
      type: object
      required: [id, questionIndex, totalQuestions, category, difficulty, question, preparationSignals]
      properties:
        id: { type: string }
        questionIndex: { type: integer }
        totalQuestions: { type: integer }
        category: { type: string }
        difficulty: { type: string }
        question: { type: string }
        preparationSignals: { type: array, items: { type: string } }
    InterviewAnswerInput:
      type: object
      required: [answer]
      properties:
        answer: { type: string, minLength: 1 }
    InterviewEvaluation:
      type: object
      required: [score, strengths, improvements, nextQuestion]
      properties:
        score: { type: integer }
        strengths: { type: array, items: { type: string } }
        improvements: { type: array, items: { type: string } }
        nextQuestion: { $ref: "#/components/schemas/InterviewSession" }