# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CareerGenie API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';
import type { ApplicationDetailScoreBreakdown } from './applicationDetailScoreBreakdown';

export type ApplicationDetail = Application & {
  summary: string;
  strengths: string[];
  gaps: string[];
  scoreBreakdown: ApplicationDetailScoreBreakdown;
  jobDescription: string;
  resumeName: string;
  acceptedCount: number;
  rejectedCount: number;
};