# lib/api-zod/src/generated/types/authConfirmationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CareerGenie API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthConfirmationResponse {
  requiresEmailConfirmation: true;
  message: string;
}