# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CareerGenie API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName?: string;
  lastName?: string;
}