# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CareerGenie API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface DashboardSummary {
  firstName: string;
  applications: number;
  interviews: number;
  offers: number;
  readiness: number;
  recentApplications: Application[];
}