# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CareerGenie API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationRisk } from './recommendationRisk';
import type { RecommendationStatus } from './recommendationStatus';

export interface Recommendation {
  id: string;
  section: string;
  originalText: string;
  suggestedText: string;
  reason: string;
  evidence: string[];
  alignment: string[];
  risk: RecommendationRisk;
  status: RecommendationStatus;
}